Разделил Меню и Фрактал
