+++ src/main/kotlin/Main.kt
@@ -2,16 +2,21 @@ import androidx.compose.desktop.ui.tooling.preview.Preview
 import androidx.compose.foundation.*
 import androidx.compose.foundation.gestures.detectDragGestures
 import androidx.compose.foundation.layout.*
-import androidx.compose.material.MaterialTheme
+import androidx.compose.material.*
 import androidx.compose.runtime.*
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.geometry.Offset
 import androidx.compose.ui.geometry.Size
 import androidx.compose.ui.graphics.*
+import androidx.compose.ui.input.key.Key.Companion.R
 import androidx.compose.ui.input.pointer.pointerInput
+import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.window.Window
 import androidx.compose.ui.window.application
+import controls.DrawingPanel
+import controls.SelectionPanel
+import controls.mainFractalWindow
 import drawing.FractalPainter
 import drawing.Painter
 import drawing.SelectionRect
@@ -20,12 +25,9 @@ import drawing.convertation.Plane
 import math.fractals.Mandelbrot
 import kotlin.math.*
 
-fun test(){
-
-}
 @Composable
 @Preview
-fun App() {
+fun App(){
     val fp = remember { FractalPainter(Mandelbrot){
         if (it == 1f) Color.Black
         else {
@@ -36,64 +38,49 @@ fun App() {
         }
     }}
     fp.plane = Plane(-2.0, 1.0, -1.0, 1.0, 0f, 0f)
-    MaterialTheme {
-        DrawingPanel(fp){size ->
-            fp.width = size.width.toInt()
-            fp.height = size.height.toInt()
-            fp.refresh = true
-        }
-        SelectionPanel{
-            fp.plane?.let{ plane ->
-                val xMin = Converter.xScr2Crt(it.topLeft.x, plane)
-                val xMax = Converter.xScr2Crt(it.topLeft.x+it.size.width, plane)
-                val yMax = Converter.yScr2Crt(it.topLeft.y, plane)
-                val yMin = Converter.yScr2Crt(it.topLeft.y+it.size.height, plane)
-                plane.xMin = xMin
-                plane.xMax = xMax
-                plane.yMin = yMin
-                plane.yMax = yMax
-                fp.refresh = true
-            }
-        }
+    MaterialTheme{
+        Scaffold(
+            topBar = {
+                // TopAppBar содержит компоненты AppBar, такие как IconButton и Text
+                TopAppBar(
+                    title = {
+                        Text(text = "SDsfasdasdasd")
+                    },
+                    navigationIcon = {
+                        // NavigationIcon представляет значок для кнопки навигации
+                        IconButton(
+                            onClick = {
+                                // Обработка события нажатия на кнопку навигации
+                            }
+                        ) {
+                        }
+                    },
+                    actions = {
+                        // Список действий, расположенных справа от заголовка
+                        IconButton(
+                            onClick = {
+                                // Обработка события нажатия на первую кнопку
+                            }
+                        ) {
+                        }
+                        IconButton(
+                            onClick = {
+                                // Обработка события нажатия на вторую кнопку
+                            }
+                        ) {
+                        }
+                    }
+                )
+            },
+            content = {
+                mainFractalWindow(fp)
+            },
+            modifier = Modifier.fillMaxSize()
+        )
     }
 }
 
-@OptIn(ExperimentalFoundationApi::class)
-@Composable
-fun SelectionPanel(
-    onSelected: (SelectionRect)->Unit
-) {
-    var rect by remember {mutableStateOf(SelectionRect(Offset.Zero))}
-    Canvas(Modifier.fillMaxSize().padding(8.dp).pointerInput(Unit){
-        detectDragGestures(
-            onDragStart = {
-                rect = SelectionRect(it)
-            },
-            onDrag = {
-                rect.addPoint(it)
-            },
-            onDragEnd = {
-                onSelected(rect)
-                rect = SelectionRect(Offset.Zero)
-            },
-            matcher = PointerMatcher.Primary)
-    }){
-        drawRect(Color(0f, 1f, 1f, 0.3f), rect.topLeft, rect.size)
-    }
-}
 
-@Composable
-fun DrawingPanel(
-    fp: Painter,
-    onResize: (Size)-> Unit = {},
-) {
-    Canvas(Modifier.fillMaxSize().padding(8.dp)) {
-        if(fp.width != size.width.toInt() || fp.height != size.height.toInt() )
-            onResize(size)
-
-        fp.paint(this)
-    }
-}
 
 
 fun main() = application {
@@ -104,3 +91,4 @@ fun main() = application {
         App()
     }
 }
+
